feat: Add Get & Sync Vendor
