+++ backend/package.json
@@ -23,6 +23,7 @@
         "@nestjs/swagger": "^5.2.1",
         "@nestjs/typeorm": "^8.1.4",
         "faker-js": "^1.0.0",
+        "knex": "^2.1.0",
         "lodash": "^4.17.21",
         "pg": "^8.7.3",
         "reflect-metadata": "^0.1.13",
