+++ backend/src/feature/plan/plan.module.ts
@@ -1,7 +1,7 @@
 import { Module, OnApplicationBootstrap } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { BigQueryService } from '../../provider/warehouse/bigquery.service';
+import { BigQueryProvider } from '../../provider/warehouse/bigquery.service';
 
 import { Plan } from './plan/plan.entity';
 import { PlanService } from './plan/plan.service';
@@ -14,15 +14,20 @@ import { PlanItemController } from './plan-item/plan-item.controller';
 import { PlanStatus } from './plan-status/plan-status.entity';
 import { PlanStatusService } from './plan-status/plan-status.service';
 
+import { Vendor } from './vendor/vendor.entity';
+import { VendorService } from './vendor/vendor.service';
+import { VendorController } from './vendor/vendor.controller';
+
 @Module({
-    imports: [TypeOrmModule.forFeature([Plan, PlanItem, PlanStatus])],
+    imports: [TypeOrmModule.forFeature([Plan, PlanItem, PlanStatus, Vendor])],
     providers: [
-        BigQueryService,
+        BigQueryProvider,
         PlanService,
         PlanItemService,
         PlanStatusService,
+        VendorService,
     ],
-    controllers: [PlanController, PlanItemController],
+    controllers: [PlanController, PlanItemController, VendorController],
 })
 export class PlanModule implements OnApplicationBootstrap {
     constructor(private planStatusService: PlanStatusService) {}
